@@ -11,28 +11,29 @@ use Illuminate\Support\Str;
 
 class ApiController extends Controller
 {
+    protected bool $_crud = true;
     /**
      * @var string
      */
-    protected $_alias = '';
+    protected string $_alias = '';
 
     /**
      * @var string
      */
-    protected $_model;
+    protected string $_model;
 
     /**
      * @var string
      */
-    protected $_resource;
+    protected string $_resource;
 
     /**
      * @var string[]
      */
-    protected $_searchFields = ['name'];
+    protected array $_searchFields = ['name'];
 
     /**
-     * @param string $model
+     * @param  string  $model
      */
     public function setModel(string $model): void
     {
@@ -40,7 +41,7 @@ class ApiController extends Controller
     }
 
     /**
-     * @param string $resource
+     * @param  string  $resource
      */
     public function setResource(string $resource): void
     {
@@ -48,7 +49,7 @@ class ApiController extends Controller
     }
 
     /**
-     * @param string[] $searchFields
+     * @param  string[]  $searchFields
      */
     public function setSearchFields(array $searchFields): void
     {
@@ -56,7 +57,7 @@ class ApiController extends Controller
     }
 
     /**
-     * @param string $alias
+     * @param  string  $alias
      */
     public function setAlias(string $alias): void
     {
@@ -69,28 +70,25 @@ class ApiController extends Controller
     public function __construct()
     {
         // Get Alias from controller name
-        $className = get_called_class();
+        $className           = get_called_class();
         $controllerClassName = Arr::last(explode('\\', $className));
-        if (empty($this->_alias)) {
-            $this->_alias = Str::snake(str_replace('Controller', '', $controllerClassName));
+        if (empty($this->_alias) && $this->_crud) {
+            $this->_alias = Str::slug(str_replace('Controller', '', $controllerClassName));
         }
         // Set Model Name use Str::studly
-        if (empty($this->_model)) {
-            $modelClass = 'App\\Models\\' . Str::studly($this->_alias);
+        if (empty($this->_model) && $this->_crud) {
+            $modelClass = 'App\\Models\\'.Str::studly($this->_alias);
             if (!class_exists($modelClass)) {
                 throw new Exception("Model Class not Found.");
             }
             $this->_model = $modelClass;
         }
         // Set Resource use Str::studly
-        if (empty($this->_resource)) {
-            $fileName = Str::studly($this->_alias) . 'Resource';
-            if (file_exists(__DIR__. '/../../Resources/' . $fileName . '.php')) {
-                $resourceClass = 'App\\Http\\Resources\\' . Str::studly($this->_alias) . 'Resource';
-            } else {
-                $resourceClass = 'App\\Http\\Resources\\DefaultResource';
+        if (empty($this->_resource) && $this->_crud) {
+            $resourceClass = 'App\\Http\\Resources\\'.Str::studly($this->_alias).'Resource';
+            if (!class_exists($resourceClass)) {
+                throw new Exception("Resource Class not Found.");
             }
-
             $this->_resource = $resourceClass;
         }
     }
@@ -98,30 +96,31 @@ class ApiController extends Controller
     public function index()
     {
         $condition = request()->query('search_text', '');
-        $perPage = request()->query('per_page', 10);
-        $sortField = request()->query('sort_field', '');
+        $perPage   = request()->query('per_page', 10);
+        $sortField = request()->query('sort_field', 'created_at');
         $sortValue = request()->query('sort_value', 'asc');
 
         $query = call_user_func([$this->_model, 'query']);
-        if (!empty($sortField)) $query->orderBy($sortField, $sortValue);
+        if (!empty($sortField)) {
+            $query->orderBy($sortField, $sortValue);
+        }
         if (!empty($condition)) {
             $searchFields = $this->_searchFields;
-            $query->where(function ($q) use($searchFields, $condition) {
+            $query->where(function ($q) use ($searchFields, $condition) {
                 foreach ($searchFields as $searchField) {
                     $q->orWhere($searchField, 'LIKE', '%'.$condition.'%');
                 }
             });
         }
-
         $query = $this->_extendIndexQuery($query);
-
         return call_user_func([$this->_resource, 'collection'], $query->paginate($perPage));
     }
 
     public function show()
     {
         $itemId = request()->{$this->_alias};
-        $item = call_user_func([$this->_model, 'find'], $itemId);
+        $item   = call_user_func([$this->_model, 'find'], $itemId);
+        if (empty($item)) return $this->responseNotFound();
         return new $this->_resource($item);
     }
 
@@ -137,12 +136,9 @@ class ApiController extends Controller
     public function update()
     {
         $itemId = request()->{$this->_alias};
-        $item = call_user_func([$this->_model, 'find'], $itemId);
+        $item   = call_user_func([$this->_model, 'find'], $itemId);
         if (empty($item)) {
-            return response([
-                'status' => 'FAIL',
-                'message' => 'Not Found'
-            ], 404);
+            return $this->responseNotFound();
         }
 
         $data = $this->_getDataUpdate($item);
@@ -155,38 +151,50 @@ class ApiController extends Controller
     /**
      * @throws Exception
      */
-    public function destroy()
+    public function destroy(): JsonResponse
     {
         $itemId = request()->{$this->_alias};
-        $item = call_user_func([$this->_model, 'find'], $itemId);
+        $item   = call_user_func([$this->_model, 'find'], $itemId);
         if (empty($item)) {
-            return response([
-                'status' => 'FAIL',
-                'message' => 'Not Found'
-            ], 404);
+            return $this->responseNotFound();
         }
-        $item->delete();
+        try {
+            $item->delete();
+        } catch (Exception $e) {
+            logger($this->_model.' '.$this->_alias.': '.$e->getMessage());
+            return $this->responseFail();
+        }
 
-        return response([
-            'status' => 'OK',
-            'message' => 'Delete complete'
-        ], 200);
+        return $this->responseSuccess('Delete Complete');
     }
 
-    public function responseFail(): JsonResponse
+    protected function responseSuccess($message = 'Success'): JsonResponse
     {
-        return response()->json([
-            'status' => 'Fail',
-            'message' => 'Fail'
-        ], 500);
+        return response()->json(
+            [
+                'status'  => 'success',
+                'message' => $message
+            ]);
     }
 
-    public function responseNotFound(): JsonResponse
+    protected function responseFail(): JsonResponse
     {
-        return response()->json([
-            'status' => 'Fail',
-            'message' => 'Not found'
-        ], 404);
+        return response()->json(
+            [
+                'status'  => 'fail',
+                'message' => 'Fail'
+            ],
+            500);
+    }
+
+    protected function responseNotFound(): JsonResponse
+    {
+        return response()->json(
+            [
+                'status'  => 'fail',
+                'message' => 'Not found'
+            ],
+            404);
     }
 
     protected function _getDataUpdate($item): array
@@ -207,7 +215,6 @@ class ApiController extends Controller
 
     protected function _afterSave($item, $action = 'store')
     {
-
     }
 
     protected function _beforeSave($data, $action = 'store')
