@@ -2,47 +2,58 @@
 
 namespace App\Http\Controllers\Api;
 
-use App\Helper\Common;
-use App\Http\Controllers\Controller;
+use App\Enums\Role;
 use App\Http\Resources\UserResource;
-use App\Mail\ResetPassword;
-use App\Models\PasswordResets;
 use App\Models\User;
-use Carbon\Carbon;
 use Exception;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\Facades\Cache;
-use Illuminate\Support\Facades\Date;
-use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Hash;
-use Illuminate\Support\Facades\Mail;
+use Illuminate\Support\Facades\Password;
+use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Str;
 
-class AuthController extends Controller
+class AuthController extends ApiController
 {
+    protected bool $_crud = false;
+
     public function login(): JsonResponse
     {
         $credentials = request(['email', 'password']);
+        $remember = request('remember');
+
+        $email = $credentials['email'];
+        $credentials['email'] = function ($query) use ($email) {
+            return $query->orWhere('email', $email)
+                ->orWhere('line_work_email', $email);
+        };
 
         if (!auth()->attempt($credentials)) {
             return response()->json(['error' => 'Unauthorized'], 401);
         }
 
-        return $this->respondWithToken(auth()->user()->createToken('services', [auth()->user()->role])->plainTextToken);
+        return $this->respondWithToken(auth()->user()->createToken(
+            'services',
+            [$this->_getAbility()],
+            $remember ? null : now()->addMonth()
+        )->plainTextToken);
     }
 
     public function logout(): JsonResponse
     {
-        $user = auth()->user();
-        if (!empty($user)) $user->currentAccessToken()->delete();
-        Auth::guard('web')->logout();
-        return response()->json(['message' => 'ログアウトしました。']);
+        if (request()->user('sanctum')) {
+            $user = request()->user('sanctum');
+            auth()->logout();
+            if (!empty($user)) $user->currentAccessToken()->delete();
+        }
+        return response()->json([
+            'status' => 'OK',
+        ]);
     }
 
     public function me(): JsonResponse
     {
-        $user = User::query()->where(['id' => auth()->user()->id])->first();
+        $user = User::query()->where(['id' => auth()->id()])->first();
         return response()->json(new UserResource($user));
     }
 
@@ -63,161 +74,112 @@ class AuthController extends Controller
         return new UserResource($user);
     }
 
-    public function checkToken()
+    public function checkToken(): UserResource|JsonResponse
     {
-        $token = request()->token;
-        if (empty($token)) {
-            return response([
-                'status' => 'FAIL',
-                'message' => 'Not Found'
-            ], 404);
-        }
-
-        $exist = User::query()->where('remember_token', '=', $token)
-            ->whereNull('email_verified_at')->first();
-        if (empty($exist)) {
-            return response([
-                'status' => 'FAIL',
-                'message' => 'Not Found'
-            ], 404);
-        }
-
-        return new UserResource($exist);
-    }
-
-    public function checkResetToken()
-    {
-        $token = request()->token;
-        if (empty($token)) {
-            return response([
-                'status' => 'FAIL',
-                'message' => 'Not Found'
-            ], 404);
-        }
-
-        $exist = PasswordResets::query()->orderBy('created_at', 'DESC')->first();
-
-        if (empty($exist) || $exist->token !== $token) {
-            return response([
-                'status' => 'FAIL',
-                'message' => 'Not Found'
-            ], 404);
-        }
-
-        $date = Carbon::parse($exist->created_at);
-        $now = Carbon::now();
-        $diff = $date->diffInDays($now);
-        if ($diff >= 2) {
-            return response([
-                'status' => 'FAIL',
-                'message' => 'Not Found'
-            ], 404);
-        }
-
-        $user = User::query()->where('email', '=', $exist->email)->first();
-        if (empty($user)) {
-            return response([
-                'status' => 'FAIL',
-                'message' => 'Not Found'
-            ], 404);
-        }
-
-        return new UserResource($user);
-    }
-
-
-    public function createPassword()
-    {
-        $token = request()->token;
-        $password = request()->password;
-        if (empty($token) || empty($password)) {
-            return response([
-                'status' => 'FAIL',
-                'message' => 'Not Found'
-            ], 404);
-        }
-
-        $exist = User::query()->where('remember_token', '=', $token)
-            ->whereNull('email_verified_at')->first();
-        if (empty($exist)) {
-            return response([
-                'status' => 'FAIL',
-                'message' => 'Not Found'
-            ], 404);
-        }
-
-        $exist->password = Hash::make($password);
-        $exist->markEmailAsVerified();
-        $exist->save();
-
-        return new UserResource($exist);
-    }
-
-    public function checkEmail()
-    {
-        $email = request()->email;
-        if (empty($email)) {
-            return response([
-                'status' => 'FAIL',
-                'message' => 'Not Found'
-            ], 404);
-        }
-
-        $exist = User::query()->where('email', '=', $email)->first();
-        if (empty($exist)) {
-            return response([
-                'status' => 'FAIL',
-                'message' => 'Not Found'
-            ], 404);
-        }
-
-        $token = Common::generateRandomString();
-        PasswordResets::create([
-            'email' => $email,
-            'token' => $token,
-            'created_at' => Date::now()
+        request()->validate([
+            'token' => 'required',
+            'email' => 'required|email',
         ]);
-        if (env('APP_ENV') !== 'local') Mail::to($exist['email'])->send(new ResetPassword($exist, $token));
+        $user = User::query()->where(['email' => request('email')])->first();
+        $exist = Password::tokenExists($user, request('token'));
 
-        return new UserResource($exist);
+        if (empty($exist)) {
+            return $this->responseNotFound();
+        }
+
+        return response()->json([
+            'status' => 'OK',
+        ]);
+    }
+
+    public function forgotPassword(): JsonResponse
+    {
+        request()->validate(['email' => 'required|email|exists:users,email']);
+        Password::sendResetLink(request()->only('email'));
+
+        return response()->json([
+            'status' => 'OK',
+        ]);
     }
 
     /**
      * @throws Exception
      */
-    public function resetPassword()
+    public function resetPassword(): JsonResponse
     {
-        $token = request()->token;
-        $password = request()->password;
-        $email = request()->email;
-        if (empty($token) || empty($password) || empty($email)) {
-            return response([
-                'status' => 'FAIL',
-                'message' => 'Not Found'
-            ], 404);
+        request()->validate([
+            'token' => 'required',
+            'email' => 'required|email',
+            'password' => 'required|min:6',
+        ]);
+
+        Password::reset(
+            request()->only('email', 'password', 'token'),
+            function (User $user, string $password) {
+                $user->forceFill([
+                    'password' => Hash::make($password)
+                ])->setRememberToken(Str::random(60));
+                $user->save();
+
+                $user->tokens()->delete();
+            }
+        );
+
+        return response()->json([
+            'status' => 'OK',
+        ]);
+    }
+
+    public function updateProfile(): UserResource
+    {
+        $user = auth()->user();
+        $data = request()->validate([
+            'name' => 'nullable',
+            'tel' => 'nullable',
+            'password' => 'nullable',
+        ]);
+        if (empty($data['password'])) {
+            unset($data['password']);
+        } else {
+            $data['password'] = Hash::make($data['password']);
+        }
+        $user->fill($data);
+
+        if (!empty(request()->file('picture'))) {
+
+            if (!empty($user->picture)) {
+                $parsed = parse_url($user->picture);
+                Storage::disk('public')->delete(str_replace('/storage/', '', $parsed['path']));
+            }
+
+            $link = Storage::disk('public')->put('avatars', request()->file('picture'));
+            $user->picture = asset('storage/' . $link);
         }
 
-        $exist = PasswordResets::query()->where('token', '=', $token)->first();
-        if (empty($exist) || $exist->email !== $email) {
-            return response([
-                'status' => 'FAIL',
-                'message' => 'Not Found'
-            ], 404);
-        }
-
-        $user = User::query()->where('email', '=', $email)->first();
-        if (empty($user)) {
-            return response([
-                'status' => 'FAIL',
-                'message' => 'Not Found'
-            ], 404);
-        }
-
-        $user->password = Hash::make($password);
         $user->save();
-        DB::table('password_resets')->where('email', $email)->delete();
 
         return new UserResource($user);
+    }
+
+    public function updatePassword(): JsonResponse
+    {
+        $user = auth()->user();
+
+        if (!Hash::check(request('oldPass'), $user->password)) {
+            return response()->json([
+                'status' => 'FAIL',
+                'message' => '古いパスワードが正しくありません。',
+            ], 400);
+        }
+        $user->password = Hash::make(request("newPass"));
+        $user->save();
+
+        $currentToken = $user->currentAccessToken();
+        $user->tokens()->whereNotIn('id', [$currentToken->id])->delete();
+
+        return response()->json([
+            'status' => 'OK',
+        ]);
     }
 
     protected function respondWithToken($token): JsonResponse
@@ -225,14 +187,16 @@ class AuthController extends Controller
         return response()->json([
             'access_token' => $token,
             'token_type' => 'bearer',
-            'detail' => [
-                'name' => auth()->user()->name,
-                'email' => auth()->user()->email,
-                'role' => auth()->user()->role,
-                'picture' => auth()->user()->picture,
-                'stamp' => auth()->user()->stamp,
-                'first_login' => auth()->user()->first_login,
-            ],
+            'detail' => new UserResource(auth()->user()),
         ]);
+    }
+
+    protected function _getAbility(): string
+    {
+        $role = auth()->user()->role;
+
+        if ($role == Role::ADMIN->value) return "admin";
+        if ($role == Role::MANAGER->value) return "manager";
+        return "user";
     }
 }
